# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BrainVault API
  version: 1.0.0
  description: Block-based note app server API.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PublicUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
          description: Login ID. This is not an email address.
        name:
          type: [string, 'null']
        avatarData:
          type: [string, 'null']
          description: PNG, JPEG, or WebP data URL used as the profile image.
        preferredLanguage:
          type: [string, 'null']
          enum: [en, ja, ko, fr, de, es, pt, null]
        defaultCollectionIcon:
          type: [string, 'null']
          maxLength: 32
          description: Emoji used for the built-in default collection.
        createdAt:
          type: string
        updatedAt:
          type: string
    RegisterInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 40
          pattern: '^[a-zA-Z0-9._-]+$'
          example: myid
        password:
          type: string
          minLength: 8
          maxLength: 128
        name:
          type: string
          maxLength: 80
        preferredLanguage:
          type: string
          enum: [en, ja, ko, fr, de, es, pt]
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 40
          pattern: '^[a-zA-Z0-9._-]+$'
          description: Login IDs are trimmed and normalized to lowercase before lookup.
          example: demo
        password:
          type: string
          minLength: 8
          maxLength: 128
          example: brainvault123
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: '#/components/schemas/PublicUser'
        token:
          type: string
    MfaMethods:
      type: object
      required: [totp, passkey]
      properties:
        totp:
          type: boolean
        passkey:
          type: boolean
    MfaRequiredResponse:
      type: object
      required: [mfaRequired, mfaToken, methods, expiresInSeconds]
      properties:
        mfaRequired:
          type: boolean
          const: true
        mfaToken:
          type: string
          description: Short-lived opaque token used only to finish this login.
        methods:
          $ref: '#/components/schemas/MfaMethods'
        expiresInSeconds:
          type: integer
          example: 300
...
    ProfileInput:
      type: object
      properties:
        name:
          type: [string, 'null']
          maxLength: 80
        avatarData:
          type: [string, 'null']
          description: PNG, JPEG, or WebP data URL. Decoded image data is limited to 512 KiB.
        preferredLanguage:
          type: [string, 'null']
          enum: [en, ja, ko, fr, de, es, pt, null]
        defaultCollectionIcon:
          type: [string, 'null']
          maxLength: 32
          description: Emoji used for the built-in default collection.
    PasswordChangeInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          maxLength: 128
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
...
paths:
...
  /api/auth/register:
    post:
      security: []
      summary: Register a new user with an ID, not an email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          description: Created user and JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/auth/login:
    post:
      security: []
      summary: Login with ID and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: User and JWT, or a temporary MFA session when two-step verification is configured
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthResponse'
                  - $ref: '#/components/schemas/MfaRequiredResponse'
...
  /api/auth/profile:
    patch:
      summary: Update the current user profile and account preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileInput'
      responses:
        '200':
          description: Updated current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/PublicUser'
  /api/auth/password:
    post:
      summary: Change the current user password after current-password verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeInput'
      responses:
        '200':
          description: Password changed
...